Run integration tests as part of script/test in non-CI environment
This allows us to run all the tests with a single command in dev.

It doesn't make sense to run the integration tests along with everything
else in CI, because we don't need to run the integration tests in PHP
7.4,8.1 & 8.2 environments, as that happens by default via the docker
setup. So doing that would just burn through GitHub action minutes. So
let's just run them once in a separate action.